// backend/src/auth/interface/auth.interface.ts
import { UserRole } from '../enums/permission.enum';

export interface JwtPayload {
  sub: string;
  email: string;
  role: UserRole;
  iat?: number;
  exp?: number;
}

export interface AuthResponse {
  access_token: string;
  user: {
    id: string;
    name: string;
    email: string;
    role: UserRole;
    profileImage?: string;
  };
}